fix debug print format type.
openbsd timeval structure defines the seconds part as time_t and not
int so we need to use a time_t format suitable everywhere.

+++ monkey/schedule.c
@@ -204,13 +204,13 @@ void monkey_schedule_list(void)
 
   gettimeofday(&tv, NULL);
 
-  LOG("schedule list at %ld:%ld", tv.tv_sec, tv.tv_usec);
+  LOG("schedule list at %lld:%ld", (long long)tv.tv_sec, tv.tv_usec);
 
   cur_nscb = schedule_list;
 
   while (cur_nscb != NULL) {
-    LOG("Schedule %p at %ld:%ld",
-	cur_nscb, cur_nscb->tv.tv_sec, cur_nscb->tv.tv_usec);
+    LOG("Schedule %p at %lld:%ld",
+	cur_nscb, (long long)cur_nscb->tv.tv_sec, cur_nscb->tv.tv_usec);
     cur_nscb = cur_nscb->next;
   }
 }
